+++ src/main/java/utils/ItemUtils.java
@@ -13,7 +13,7 @@ public class ItemUtils {
     }
     public static void displayGroupedItems(List<Item> items) {
         Map<String, Integer> grouped = new LinkedHashMap<>();
-        Map<String, Integer> prices = new HashMap<>();
+        Map<String, Double> prices = new HashMap<>();
 
         for (Item item : items) {
             grouped.put(item.getName(), grouped.getOrDefault(item.getName(), 0) + 1);
